Remove deborphan
deborphan was removed from Debian Sid in Bug 1065312; it is not
present in Debian Trixie and higher (except that it still exists
in Sid for riscv64 for... some reason).

Removing deborphan from Kicksecure dependencies fixes issues
when distro-morphing Debian Trixie and higher to Kicksecure. It
is also necessary for installing Kicksecure on ports-only
architectures such as ppc64.

Debian bug 1074134 (asking for suggested alternatives) remains
open, but it doesn't look like it's going to be fixed anytime
soon, so I think we should cut our losses and remove it from
Kicksecure now, and maybe add an alternative later once one
exists.